moved ssl startup check from general to specific install script

+++ install-elknode.sh
@@ -1,5 +1,27 @@
 #!/bin/bash
 
+## SSL CHECK ------------------------------------------------------------------------------------
+echo "SSL SETUP CHECK: STARTING SSL SETUP CHECK"
+
+# check ssl setup: test if truststore exists
+if [ ! -f $ELKINSTALLDIR/ssl/truststore.jks ] 
+then
+	echo "SSL SETUP CHECK: ERROR: The required file \"ssl/truststore.jks\" does not exist."
+	echo "Please run \"prepare-ssl.sh\" before booting any vagrant boxes or insert your own jks files!"
+	exit;
+fi
+
+# check ssl setup: test if keystore exists
+if [ ! -f $ELKINSTALLDIR/ssl/$(hostname)-keystore.jks ] 
+then
+	echo "SSL SETUP CHECK: ERROR: The required file \"ssl/$(hostname)-keystore.jks\" does not exist."
+	echo "Please run \"prepare-ssl.sh\" before booting any vagrant boxes or insert your own jks files!"
+	exit;
+fi
+echo "SSL SETUP CHECK: CHECK FINISHED. NO PROBLEMS DETECTED"
+## SSL CHECK END --------------------------------------------------------------------------------
+
+
 # install the required puppet modules and dependencies
 sudo puppet module install elasticsearch-elasticsearch
